[ESIMD][NFC] Reshuffle definitions to reduce dependencies in headers.
