+++ sycl/include/sycl/ext/intel/esimd/detail/types.hpp
@@ -13,7 +13,9 @@
 #include <sycl/detail/defines.hpp>
 #include <sycl/detail/stl_type_traits.hpp> // to define C++14,17 extensions
 #include <sycl/ext/intel/esimd/common.hpp>
+#include <sycl/ext/intel/esimd/detail/elem_type_traits.hpp>
 #include <sycl/ext/intel/esimd/detail/region.hpp>
+#include <sycl/ext/intel/esimd/detail/types_elementary.hpp>
 #include <sycl/half_type.hpp>
 
 #if defined(__ESIMD_DBG_HOST) && !defined(__SYCL_DEVICE_ONLY__)
@@ -36,14 +38,6 @@ template <typename BaseTy, typename RegionTy> class simd_view;
 
 namespace detail {
 
-template <int N>
-using uint_type_t = std::conditional_t<
-    N == 1, uint8_t,
-    std::conditional_t<
-        N == 2, uint16_t,
-        std::conditional_t<N == 4, uint32_t,
-                           std::conditional_t<N == 8, uint64_t, void>>>>;
-
 // forward declarations of major internal simd classes
 template <typename Ty, int N> class simd_mask_impl;
 template <typename RawT, int N, class Derived, class SFINAE = void>
@@ -56,83 +50,12 @@ class simd_obj_impl;
 using simd_mask_elem_type = unsigned short;
 template <int N> using simd_mask_type = simd_mask_impl<simd_mask_elem_type, N>;
 
-// @{
-// Checks if given type T is a raw clang vector type, plus provides some info
-// about it if it is.
-
-struct invalid_element_type;
-
-template <class T> struct is_clang_vector_type : std::false_type {
-  static inline constexpr int length = 0;
-  using element_type = invalid_element_type;
-};
-
-template <class T, int N>
-struct is_clang_vector_type<T __attribute__((ext_vector_type(N)))>
-    : std::true_type {
-  static inline constexpr int length = N;
-  using element_type = T;
-};
-template <class T>
-static inline constexpr bool is_clang_vector_type_v =
-    is_clang_vector_type<T>::value;
-
-// @}
-
-template <typename T>
-using remove_cvref_t =
-    sycl::detail::remove_cv_t<sycl::detail::remove_reference_t<T>>;
-
-// is_esimd_arithmetic_type
-template <class...> struct make_esimd_void {
-  using type = void;
-};
-template <class... Tys>
-using __esimd_void_t = typename make_esimd_void<Tys...>::type;
-
-template <class Ty, class = void>
-struct is_esimd_arithmetic_type : std::false_type {};
-
-template <class Ty>
-struct is_esimd_arithmetic_type<
-    Ty, __esimd_void_t<std::enable_if_t<std::is_arithmetic_v<Ty>>,
-                       decltype(std::declval<Ty>() + std::declval<Ty>()),
-                       decltype(std::declval<Ty>() - std::declval<Ty>()),
-                       decltype(std::declval<Ty>() * std::declval<Ty>()),
-                       decltype(std::declval<Ty>() / std::declval<Ty>())>>
-    : std::true_type {};
-
-template <typename Ty>
-static inline constexpr bool is_esimd_arithmetic_type_v =
-    is_esimd_arithmetic_type<Ty>::value;
-
-// is_vectorizable_type
-template <typename Ty>
-struct is_vectorizable : std::conditional_t<is_esimd_arithmetic_type_v<Ty>,
-                                            std::true_type, std::false_type> {};
-
-template <typename Ty>
-static inline constexpr bool is_vectorizable_v = is_vectorizable<Ty>::value;
-
 template <typename T>
 static inline constexpr bool is_esimd_scalar_v =
     sycl::detail::is_arithmetic<T>::value;
 
 template <typename T>
 using is_esimd_scalar = typename std::bool_constant<is_esimd_scalar_v<T>>;
-
-// raw_vector_type, using clang vector type extension.
-template <typename Ty, int N> struct raw_vector_type {
-  static_assert(!std::is_const<Ty>::value, "const element type not supported");
-  static_assert(is_vectorizable_v<Ty>, "element type not supported");
-  static_assert(N > 0, "zero-element vector not supported");
-
-  static constexpr int length = N;
-  using type = Ty __attribute__((ext_vector_type(N)));
-};
-
-template <typename Ty, int N>
-using vector_type_t = typename raw_vector_type<Ty, N>::type;
 
 // @{
 // Checks if given type T derives from simd_obj_impl or is equal to it.
@@ -143,12 +66,6 @@ struct is_simd_obj_impl_derivative : public std::false_type {};
 template <typename RawT, int N, class Derived>
 struct is_simd_obj_impl_derivative<simd_obj_impl<RawT, N, Derived>>
     : public std::true_type {};
-
-template <class T, class SFINAE = void> struct element_type_traits;
-template <class T>
-using __raw_t = typename __ESIMD_DNS::element_type_traits<T>::RawT;
-template <class T>
-using __cpp_t = typename __ESIMD_DNS::element_type_traits<T>::EnclosingCppT;
 
 // Specialization for all other types.
 template <typename T, int N, template <typename, int> class Derived>
@@ -352,15 +269,6 @@ std::enable_if_t<is_clang_vector_type_v<To> && is_clang_vector_type_v<From>, To>
   }
 }
 
-/// Base case for checking if a type U is one of the types.
-template <typename U> constexpr bool is_type() { return false; }
-
-template <typename U, typename T, typename... Ts> constexpr bool is_type() {
-  using UU = typename std::remove_const_t<U>;
-  using TT = typename std::remove_const_t<T>;
-  return std::is_same<UU, TT>::value || is_type<UU, Ts...>();
-}
-
 // calculates the number of elements in "To" type
 template <typename ToEltTy, typename FromEltTy, int FromN,
           typename = std::enable_if_t<is_vectorizable<ToEltTy>::value>>
@@ -390,6 +298,91 @@ bitcast(vector_type_t<FromEltTy, FromN> Val) {
   using VTy = vector_type_t<ToEltTy, ToN>;
   return reinterpret_cast<VTy>(Val);
 }
+
+// Get computation type of a binary operator given its operand types:
+// - if both types are arithmetic - return CPP's "common real type" of the
+//   computation (matches C++)
+// - if both types are simd types, they must be of the same length N,
+//   and the returned type is simd<T, N>, where N is the "common real type" of
+//   the element type of the operands (diverges from clang)
+// - otherwise, one type is simd and another is arithmetic - the simd type is
+//   returned (matches clang)
+
+struct invalid_computation_type;
+
+template <class T1, class T2, class SFINAE = void> struct computation_type {
+  using type = invalid_computation_type;
+};
+
+template <class T1, class T2>
+struct computation_type<T1, T2,
+                        std::enable_if_t<is_valid_simd_elem_type_v<T1> &&
+                                         is_valid_simd_elem_type_v<T2>>> {
+private:
+  template <class T> using tr = element_type_traits<T>;
+  template <class T>
+  using native_t =
+      std::conditional_t<tr<T>::use_native_cpp_ops, typename tr<T>::RawT,
+                         typename tr<T>::EnclosingCppT>;
+  static inline constexpr bool is_wr1 = is_wrapper_elem_type_v<T1>;
+  static inline constexpr bool is_wr2 = is_wrapper_elem_type_v<T2>;
+  static inline constexpr bool is_fp1 = is_generic_floating_point_v<T1>;
+  static inline constexpr bool is_fp2 = is_generic_floating_point_v<T2>;
+
+public:
+  using type = std::conditional_t<
+      !is_wr1 && !is_wr2,
+      // T1 and T2 are both std C++ types - use std C++ type promotion
+      decltype(std::declval<T1>() + std::declval<T2>()),
+      std::conditional_t<
+          std::is_same_v<T1, T2>,
+          // Types are the same wrapper type - return any
+          T1,
+          std::conditional_t<is_fp1 != is_fp2,
+                             // One of the types is floating-point - return it
+                             // (e.g. computation_type<int, sycl::half> will
+                             // yield sycl::half)
+                             std::conditional_t<is_fp1, T1, T2>,
+                             // both are either floating point or integral -
+                             // return result of C++ promotion of the native
+                             // types
+                             decltype(std::declval<native_t<T1>>() +
+                                      std::declval<native_t<T2>>())>>>;
+};
+
+template <class T1, class T2>
+struct computation_type<
+    T1, T2,
+    std::enable_if_t<is_simd_like_type_v<T1> || is_simd_like_type_v<T2>>> {
+private:
+  using Ty1 = element_type_t<T1>;
+  using Ty2 = element_type_t<T2>;
+  using EltTy = typename computation_type<Ty1, Ty2>::type;
+
+  static constexpr int N1 = [] {
+    if constexpr (is_simd_like_type_v<T1>) {
+      return T1::length;
+    } else {
+      return 0;
+    }
+  }();
+  static constexpr int N2 = [] {
+    if constexpr (is_simd_like_type_v<T2>) {
+      return T2::length;
+    } else {
+      return 0;
+    }
+  }();
+  static_assert((N1 == N2) || (N1 == 0) || (N2 == 0), "size mismatch");
+  static constexpr int N = N1 ? N1 : N2;
+
+public:
+  using type = simd<EltTy, N>;
+};
+
+template <class T1, class T2 = T1>
+using computation_type_t =
+    typename computation_type<remove_cvref_t<T1>, remove_cvref_t<T2>>::type;
 
 } // namespace detail
